feat: recommend no sudo
Intruduces new result type: recommendation

A recommendation is neither a success nor failure, but a common practice
to avoid although there are some valid use cases. It should not result
in a failure in any execution contexts.

First recommendation is to avoid sudo/su and direct users to gosu if
neccessary to elevate privileges within the Dockerfile.