feat: support config of subset of default rules
This introduces skip_default_rules (bool) and include_rules ([]string)
configuration options. These options allow the user to include only a
small set of the default rules. This simplifies, for example, if the
user wants only a list of custom regex-based rules. It also simplifies
the example tests which validate output as they no longer need to be
updated as new rules are added.